requirements.txt
discord.py==2.0.1
pandas==1.4.2
tweepy==4.12.1
